requirements.txt
fastapi==0.111.0
psycopg2-binary==2.9.9
sqlmodel==0.0.19
alembic==1.13.2
uvicorn==0.30.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==0.19.0

httpx
python-multipart






